+++ rtsp/lib/rtsp/connection_manager.ex
@@ -340,7 +340,7 @@ defmodule Membrane.RTC.Engine.Endpoint.RTSP.ConnectionManager do
   end
 
   defp get_attribute(video_attributes, attribute, default \\ nil) do
-    case ExSDP.Media.get_attribute(video_attributes, attribute) do
+    case ExSDP.get_attribute(video_attributes, attribute) do
       {^attribute, value} -> value
       %^attribute{} = value -> value
       _other -> default