+++ webrtc/lib/endpoint_bin/extension/mid.ex
@@ -6,7 +6,6 @@ defmodule Membrane.WebRTC.Extension.Mid do
   """
   @behaviour Membrane.WebRTC.Extension
   alias ExSDP.Attribute.Extmap
-  alias ExSDP.Media
   alias Membrane.WebRTC.Extension
 
   @name :mid
@@ -32,7 +31,7 @@ defmodule Membrane.WebRTC.Extension.Mid do
 
   @impl true
   def add_to_media(media, id, _direction, _payload_types),
-    do: Media.add_attribute(media, %Extmap{id: id, uri: @uri})
+    do: ExSDP.add_attribute(media, %Extmap{id: id, uri: @uri})
 
   @impl true
   def uri, do: @uri