+++ sip/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.4.0-dev
+* Update deps [#393](https://github.com/fishjam-dev/membrane_rtc_engine/pull/393)
 
 ## 0.3.0
 * Update deps [#374](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/374)